Add rules for building lexical nodes
A rule is either an 'and', 'or' or a constant rule. 'and' and 'or'
rules contain other rules while constant rules cannot contain other
rules. With these primitives we can build other complex rules.

To perform lexical analysis we need to be able to transform these rules
into a tree of lexical nodes that represent each of the states that
the lexer can transition for a given rule.

At this stage I haven't added transitions only nodes.