Update Compiler/Clr assembly to .NET core
Update the CLR compiler assembly to use .NET Core.

This change has some far reaching consequences because .NET Core does not yet
support saving dynamically generated assemblies to disk.

This means that we can no longer actually create compiled versions of
slang files, rather slang has to operate as an interpreter creating the
dynamic assemblies on the fly as it compiles them to IL.

It might be the case that the slang compiler executable should simply
interpret libraries and load them into a REPL environment and contain a
switch to allow you to run the script without the REPL.