Update Compiler/Coco project to use .NET Core
Update the compiler Coco project to use .NET Core.

In .NET Core streams no longer implement the 'Close' method so in this
case I have had to update the buffer to call 'Dispose' instead which
does the same thing. Ideally the Buffer class would implement
IDisposable but this code is auto-generated by Coco.exe so it is not
possible at this stage.

At the moment this assembly compiles correctly however it will not be
updated if you were to make changes to the .atg file. This is because
the custom build step to call the executable is missing.