Add a script to generate scanner and parser from grammar
Moving to .NET core forced me to remove the scanner and grammar
generation step from the MSBuild .csproj file.

I've added these back in again but I have had to include a step to fix
up code that is broken during the migration, i.e. the deprecation of the
'System.IO.Stream.Close' method.

This is far from perfect as the resulting file is missing the BOM which
can cause problems in Microsoft editors that expect this to be present.