Use xUnit rather than NUnit
The xUnit test runner is supported by .NET Core pretty much out of the
box. Whereas the NUnit test runner appeared to work with the
project.json 'testRunner' property but now requires the package to be
added an run implicitly presumably.

I didn't have the time and patience to try to figure out how to make the
NUnit runner work, instead it seemed easier to just convert these tests
to use xUnit. Mostly the code in the tests uses FluentAssertions rather
than NUnit directly so it's remarkably easy to convert them.