Stop using IL generation
It turns out that using IL generation under .NET Core is not very easy
at all. When I try to generate types using the assembly builder I'm not
able to actually load the types. Dynamic assemblies cannot be saved to
disk either which creates problems because we can't compile our Slang
code to an executable or library.

Therefore I've made the decision to remove support here for generating
IL until such time as it is reasonable to do it on .NET Core.