Add note about using properties rather than functions
While it is appealing to use functions for values on a type at this
stage we are more likely to want to use properties.

This is because a type will be a class that contains properties and may
have various constructors.

At least this is my thinking at the moment.