Generate types as nested classes
Allow us to generate a nested class that corresponds to a type
definition. This gives us modules, functions, and types as our basic
high level building blocks.

Nested classes are preferrable here so that any classes that are
required to implement special features of the type are hidden by the
module and can be shared with other exported types.